LayoutTests/http/tests/security/xssAuditor/object-tag-javascript-url-expected.txt

CONSOLE MESSAGE: line 1: Refused to load an object. URL found within request: "javascript:alert(document.domain)".

CONSOLE MESSAGE: line 1: Refused to load an object. URL found within request: "javascript:alert(document.domain)".
